requirements.txt
requests==2.22.0
python-dotenv==0.12.0
chart-studio==1.1.0 